requirements_test.txt
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-homeassistant-custom-component==0.13.76
pytest-cov==4.1.0
pytest-timeout==2.2.0
